[refactor] gan.sample now returns a BatchSample
